chore(router): move RouterIntent out of interface
