+++ core/src/components/router/utils/constants.ts
@@ -1,0 +1,6 @@
+
+export const enum RouterIntent {
+  None = 0,
+  Forward = 1,
+  Back = -1,
+}
