+++ core/src/components/router/utils/dom.ts
@@ -1,5 +1,5 @@
 import { NavOutletElement, RouteChain, RouteID } from '../../../interface';
-import { RouterIntent } from './interface';
+import { RouterIntent } from './constants';
 
 export async function writeNavState(root: HTMLElement | undefined, chain: RouteChain, intent: RouterIntent, index: number, changed = false): Promise<boolean> {
   try {
